% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{as.multivar}
\alias{as.multivar}
\alias{multivar}
\title{as.multivar}
\usage{
as.multivar(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{parameters of \code{\link[=subset.univar]{subset.univar()}}}
}
\description{
as.multivar
}
\examples{
library(ranef.tzh)
library(ordinal)
m1 = datasets::mtcars |> 
 within.data.frame(expr = {
  cyl = factor(cyl)
  gear = factor(gear)
  vs = ifelse(as.logical(vs), yes = 'straight', no = 'V-shaped')
  am = ifelse(as.logical(am), yes = 'manual', no = 'automatic')
 }) |>
 lm(mpg ~ cyl + am + hp + wt + qsec + drat + disp, data = _)
m2 = lme4::glmer(outcome ~ treatment + visit + (1|patientID), data = lme4::toenail,
  family = binomial, nAGQ = 20)
m3 = clmm(SURENESS ~ PROD + SOUPTYPE + (1|RESP) + (1|RESP:PROD), data = soup,
 link = 'probit', threshold = 'equidistant')
  
library(ecip); list(
 'univar, `lm`' = m1 |> as.univar(),
 'multivar, `lm`' = m1 |> as.univar() |> as.multivar(subset = min_pvalue < .1),
 'multivar, `merMod`' = m2 |> as.univar() |> as.multivar(subset = min_pvalue < .1)# ,
 # 'multivar, `clmm`' = m3 |> as.univar() |> as.multivar(subset = min_pvalue < .1)# still bug
) |> fastmd::render2html()


}
\keyword{internal}
