% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_over_dispersion.R
\name{check_over_dispersion}
\alias{check_over_dispersion}
\title{check_over_dispersion}
\usage{
check_over_dispersion(model, ...)
}
\arguments{
\item{model}{a \link[stats]{glm} model for \link[stats]{poisson} family}

\item{...}{..}
}
\description{
select pois vs. negbin based on overdispersion
}
\examples{
glm(Days ~ ., family = poisson, data = MASS::quine) |>
 check_over_dispersion() |> 
 stepAIC_complete()

}
