% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.nb.R
\name{getCanonicalLink.negbin}
\alias{getCanonicalLink.negbin}
\title{getCanonicalLink}
\usage{
\method{getCanonicalLink}{negbin}(x)
}
\arguments{
\item{x}{object returned from function \link[MASS]{glm.nb}}
}
\description{
..
}
\examples{
library(MASS)
glm.nb(Days ~ Sex/(Age + Eth*Lrn), data = quine) |>
  getCanonicalLink.negbin()

}
\keyword{internal}
