% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{md_.multivar}
\alias{md_.multivar}
\title{Model Description of \link{multivar} Object}
\usage{
\method{md_}{multivar}(x, xnm, ...)
}
\arguments{
\item{x}{a \link{multivar} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
library(lme4.tzh)
library(HSAUR3)
library(ordinal)
m1 = lm(mpg ~ cyl + am + hp + wt + qsec + drat + disp, data = ecip::mtc)
m2 = glmer(outcome ~ treatment + visit + (1|patientID), data = toenail,
  family = binomial, nAGQ = 20)
m3 = clmm(SURENESS ~ PROD + SOUPTYPE + (1|RESP) + (1|RESP:PROD), data = soup,
 link = 'probit', threshold = 'equidistant')
  
library(ecip); list(
 'univar, `lm`' = m1 |> as.univar(),
 'multivar, `lm`' = m1 |> as.univar() |> as.multivar(subset = min_pvalue < .1),
 'multivar, `merMod`' = m2 |> as.univar() |> as.multivar(subset = min_pvalue < .1)# ,
 # 'multivar, `clmm`' = m3 |> as.univar() |> as.multivar(subset = min_pvalue < .1)# still bug
) |> fastmd::render2html()
}
\keyword{internal}
