% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{md_.rlm}
\alias{md_.rlm}
\title{R Markdown Lines for \link[MASS]{rlm}}
\usage{
\method{md_}{rlm}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[MASS]{rlm}
}
\examples{
library(rmd.tzh); library(ecip); list(
 '`rlm`' = rlm(stack.loss ~ ., data = stackloss)
) |> render_(file = 'rlm')

}
\keyword{internal}
