% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_complete.R
\name{md_.stepAIC}
\alias{md_.stepAIC}
\title{R Markdown Lines for \code{stepAIC}}
\usage{
\method{md_}{stepAIC}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \code{stepAIC}
}
\keyword{internal}
