% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{md_.univar}
\alias{md_.univar}
\title{R Markdown Lines for \link{univar} Object}
\usage{
\method{md_}{univar}(x, xnm, ...)
}
\arguments{
\item{x}{\link{univar} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
R Markdown Lines for \link{univar} Object
}
\keyword{internal}
