% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.nb.R
\name{negbin}
\alias{negbin}
\title{Support for \link[MASS]{glm.nb}}
\description{
Support for \link[MASS]{glm.nb}
}
\examples{
library(MASS)
nb = glm.nb(Days ~ Sex/(Age + Eth*Lrn), data = quine)
class(nb) # c('negbin', 'glm', 'lm')

library(ecip)
nb |> getCanonicalLink()
list(negbin = nb) |>
  fastmd::render2html()

}
\keyword{internal}
