% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polr.R
\name{polr}
\alias{polr}
\title{Support for \link[MASS]{polr}}
\description{
Support for \link[MASS]{polr}
}
\examples{
library(MASS)
library(ecip)
options(contrasts = c('contr.treatment', 'contr.poly'))
list('`polr`' = polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing, Hess = TRUE)) |>
 fastmd::render2html()
}
\keyword{internal}
