% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{rlm}
\alias{rlm}
\title{Support for \link[MASS]{rlm}}
\description{
Support for \link[MASS]{rlm}
}
\examples{
library(ecip); list(
 '`rlm`' = rlm(stack.loss ~ ., data = stackloss) # c('rlm', 'lm')
) |> fastmd::render2html()
}
\keyword{internal}
