% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{S3_multivar}
\alias{S3_multivar}
\alias{endpoint.multivar}
\alias{nobsText.multivar}
\alias{desc_.multivar}
\title{Additional S3 methods for \code{multivar}}
\usage{
\method{endpoint}{multivar}(x)

\method{nobsText}{multivar}(x)

\method{desc_}{multivar}(x)
}
\arguments{
\item{x}{\code{multivar}}
}
\description{
Additional S3 methods for \code{multivar}
}
\keyword{internal}
