% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{S3_rlm}
\alias{S3_rlm}
\alias{desc_.rlm}
\alias{.pval.rlm}
\alias{.pval.summary.rlm}
\title{S3 methods for \link[MASS]{rlm}}
\usage{
\method{desc_}{rlm}(x)

\method{.pval}{rlm}(x)

\method{.pval}{summary.rlm}(x)
}
\arguments{
\item{x}{an \link[MASS]{rlm} object}
}
\description{
S3 methods for \link[MASS]{rlm}
}
\examples{
library(ecip); list(
 '`rlm`' = rlm(stack.loss ~ ., data = stackloss)
) |> fastmd::render2html()
}
\keyword{internal}
