% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_complete.R
\name{Sprintf.stepAIC}
\alias{Sprintf.stepAIC}
\title{Sprintf.stepAIC}
\usage{
Sprintf.stepAIC(x)
}
\arguments{
\item{x}{returned object of \code{\link[=stepAIC_complete]{stepAIC_complete()}}}
}
\description{
Sprintf.stepAIC
}
