% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC4.R
\name{stepAIC4}
\alias{stepAIC4}
\title{stepAIC4}
\usage{
stepAIC4(
  object,
  scope,
  scale = 0,
  direction = c("both", "backward", "forward"),
  trace = 1,
  keep = NULL,
  steps = 1000,
  use.start = FALSE,
  k = 2,
  ...
)
}
\arguments{
\item{object, scope, scale, direction, trace, keep, steps, use.start, k, ...}{see explanation in function \link[MASS]{stepAIC}.}
}
\description{
Stepwise variable selection by AIC, for S4 regression object (e.g. \link[lme4]{glmer}).
}
\note{
Function \code{\link[=stepAIC4]{stepAIC4()}} is copied from \link[MASS]{stepAIC}, with as little change as possible.
}
\examples{
library(lme4); library(HSAUR3)
?lme4::glmer
gm2 = glmer(outcome ~ treatment*visit + (1|patientID), data=toenail, family=binomial,nAGQ=20)
gm2 |> stepAIC4(direction = 'backward')
}
