% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_complete.R
\name{stepAIC_complete}
\alias{stepAIC_complete}
\title{Perform \link[MASS]{stepAIC} on Complete Data}
\usage{
stepAIC_complete(object)
}
\arguments{
\item{object}{a regression model}
}
\description{
..
}
\note{
\link[MASS]{stepAIC} (as of 2025-02-19) ??? does not have a parameter for the end user to specify the
threshold of AIC improvement, that should be regarded as an improvement.
Instead, this quantity is hard coded (in \code{if (bAIC >= AIC + 1e-07) break}).
}
\examples{
stk = stackloss
stk[2L, 'Air.Flow'] = stk[5L, 'Water.Temp'] = 
  stk[16L, 'Acid.Conc.'] = stk[21L, 'stack.loss'] = NA

lm(stack.loss ~ ., data = stk) |> stepAIC_complete() |> summary()
lm(stack.loss ~ Air.Flow, data = stk) |> stepAIC_complete() |> summary()

mtc = mtcars |> within.data.frame(expr = {
 cyl = factor(cyl)
 gear = factor(gear)
 vs = as.logical(vs)
 am = as.logical(am)
})
glm(am ~ hp + wt + vs, data = mtc, family = binomial) |> stepAIC_complete()
glm(am ~ ., family = binomial, data = mtc) |> stepAIC_complete() |> suppressWarnings()

library(MASS)
glm.nb(Days ~ ., data = quine) |> stepAIC_complete()

library(ordinal)
clm(rating ~ temp * contact, data = ordinal::wine) |>
 stepAIC_complete()

}
