% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{MASS_bib}
\alias{MASS_bib}
\alias{.akaike74}
\title{\link[utils]{bibentry} for \CRANpkg{MASS}}
\usage{
.akaike74(key = "Akaike74", ...)
}
\arguments{
\item{key, ...}{\link[utils]{bibentry}}
}
\description{
\link[utils]{bibentry} for \CRANpkg{MASS}
}
\keyword{internal}
