% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polr.R
\name{S3_polr}
\alias{S3_polr}
\alias{getLink.polr}
\alias{desc_.polr}
\alias{.pval.summary.polr}
\title{S3 methods for \link[MASS]{polr}}
\usage{
\method{getLink}{polr}(x)

\method{desc_}{polr}(x)

\method{.pval}{summary.polr}(x)
}
\arguments{
\item{x}{a \link[MASS]{polr} object}
}
\description{
S3 methods for \link[MASS]{polr}
}
\examples{
library(MASS)
library(ecip)
options(contrasts = c('contr.treatment', 'contr.poly'))
m = polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing, Hess = TRUE)
.pval(m)
}
\keyword{internal}
